main: change the default wipe strategy
Make the more aggressive and destructive write strategy default. This
trades reliability for performance. If this tool ever goes out of
engineering hands, this patch has to be reverted.

As another reason for this patch, the wipe_by_biggest_erases method is
far more tested than the wipe_if_changes one.

Signed-off-by: Robert Jarzmik <robert.jarzmik@free.fr>